Use enums header file
