+++ keyboards/planck/keymaps/prog_qgmlwb/src/enums.h
@@ -1,0 +1,26 @@
+// Copyright 2022 David Kristoffersen (@davidkristoffersen)
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+// Layouts enum
+enum keyboard_layouts {
+    QGMLWB_NO,
+    QGMLWB_US,
+
+    QWERTY,
+    COLEMAK,
+    DVORAK,
+
+    NORMAL_LOWER,
+    NORMAL_RAISE,
+
+    QGMLWB_NO_LOWER,
+    QGMLWB_US_LOWER,
+
+    QGMLWB_NO_RAISE,
+    QGMLWB_US_RAISE,
+
+    ADJUST,
+    NUMPAD,
+};
